+++ nbconvert/__init__.py
@@ -21,6 +21,7 @@ try:
         ScriptExporter,
         SlidesExporter,
         TemplateExporter,
+        WebHTMLExporter,
         WebPDFExporter,
         export,
         get_export_names,
@@ -48,6 +49,7 @@ __all__ = [
     "ScriptExporter",
     "SlidesExporter",
     "TemplateExporter",
+    "WebHTMLExporter",
     "WebPDFExporter",
     "__version__",
     "export",
