+++ nbconvert/exporters/webpdf.py
@@ -3,35 +3,21 @@
 # Copyright (c) IPython Development Team.
 # Distributed under the terms of the Modified BSD License.
 
-import asyncio
-import concurrent.futures
-import os
-import subprocess
-import sys
-import tempfile
-from importlib import util as importlib_util
+from traitlets import Bool, Int, default
 
-from traitlets import Bool, Int, List, Unicode, default
+from .webhtml import PLAYWRIGHT_INSTALLED, WebHTMLExporter  # noqa: F401
 
-from .html import HTMLExporter
-
-PLAYWRIGHT_INSTALLED = importlib_util.find_spec("playwright") is not None
-IS_WINDOWS = os.name == "nt"
+# NOTE: reexport PLAYWRIGHT_INSTALLED from webhtml for backwards compatibility
 
 
-class WebPDFExporter(HTMLExporter):
+class WebPDFExporter(WebHTMLExporter):
     """Writer designed to write to PDF files.
 
-    This inherits from :class:`HTMLExporter`. It creates the HTML using the
+    This inherits from :class:`WebHTMLExporter`. It creates the HTML using the
     template machinery, and then run playwright to create a pdf.
     """
 
     export_from_notebook = "PDF via HTML"
-
-    allow_chromium_download = Bool(
-        False,
-        help="Whether to allow downloading Chromium if no suitable version is found on the system.",
-    ).tag(config=True)
 
     paginate = Bool(
         True,
@@ -66,79 +52,11 @@ class WebPDFExporter(HTMLExporter):
     def _template_name_default(self):
         return "webpdf"
 
-    disable_sandbox = Bool(
-        False,
-        help="""
-        Disable chromium security sandbox when converting to PDF.
-
-        WARNING: This could cause arbitrary code execution in specific circumstances,
-        where JS in your notebook can execute serverside code! Please use with
-        caution.
-
-        ``https://github.com/puppeteer/puppeteer/blob/main@%7B2020-12-14T17:22:24Z%7D/docs/troubleshooting.md#setting-up-chrome-linux-sandbox``
-        has more information.
-
-        This is required for webpdf to work inside most container environments.
-        """,
-    ).tag(config=True)
-
-    browser_args = List(
-        Unicode(),
-        help="""
-        Additional arguments to pass to the browser rendering to PDF.
-
-        These arguments will be passed directly to the browser launch method
-        and can be used to customize browser behavior beyond the default settings.
-        """,
-    ).tag(config=True)
-
-    def run_playwright(self, html):
+    def run_playwright(self, html, _postprocess=None):
         """Run playwright."""
 
-        async def main(temp_file):
-            """Run main playwright script."""
-
-            try:
-                from playwright.async_api import (  # type: ignore[import-not-found] # noqa: PLC0415,
-                    async_playwright,
-                )
-            except ModuleNotFoundError as e:
-                msg = (
-                    "Playwright is not installed to support Web PDF conversion. "
-                    "Please install `nbconvert[webpdf]` to enable."
-                )
-                raise RuntimeError(msg) from e
-
-            if self.allow_chromium_download:
-                cmd = [sys.executable, "-m", "playwright", "install", "chromium"]
-                subprocess.check_call(cmd)  # noqa: S603
-
-            playwright = await async_playwright().start()
-            chromium = playwright.chromium
-
-            args = self.browser_args
-            if self.disable_sandbox:
-                args.append("--no-sandbox")
-
-            try:
-                browser = await chromium.launch(
-                    handle_sigint=False, handle_sigterm=False, handle_sighup=False, args=args
-                )
-            except Exception as e:
-                msg = (
-                    "No suitable chromium executable found on the system. "
-                    "Please use '--allow-chromium-download' to allow downloading one,"
-                    "or install it using `playwright install chromium`."
-                )
-                await playwright.stop()
-                raise RuntimeError(msg) from e
-
-            page = await browser.new_page()
-            await page.emulate_media(media="print")
-            await page.wait_for_timeout(100)
-            await page.goto(f"file://{temp_file.name}", wait_until="networkidle")
-            await page.wait_for_timeout(self.page_render_timeout)
-
+        async def _pdf_postprocess(page, browser, playwright):
+            """Post-process the PDF output."""
             pdf_params = {"print_background": True}
             if not self.paginate:
                 # Floating point precision errors cause the printed
@@ -161,29 +79,9 @@ class WebPDFExporter(HTMLExporter):
                         "height": min(height, 200 * 72),
                     }
                 )
-            pdf_data = await page.pdf(**pdf_params)
+            return await page.pdf(**pdf_params)
 
-            await browser.close()
-            await playwright.stop()
-            return pdf_data
-
-        pool = concurrent.futures.ThreadPoolExecutor()
-        # Create a temporary file to pass the HTML code to Chromium:
-        # Unfortunately, tempfile on Windows does not allow for an already open
-        # file to be opened by a separate process. So we must close it first
-        # before calling Chromium. We also specify delete=False to ensure the
-        # file is not deleted after closing (the default behavior).
-        temp_file = tempfile.NamedTemporaryFile(  # noqa: SIM115
-            suffix=".html", delete=False
-        )
-        with temp_file:
-            temp_file.write(html.encode("utf-8"))
-        try:
-            pdf_data = pool.submit(asyncio.run, main(temp_file)).result()
-        finally:
-            # Ensure the file is deleted even if playwright raises an exception
-            os.unlink(temp_file.name)
-        return pdf_data
+        return super().run_playwright(html, _postprocess=_pdf_postprocess)
 
     def from_notebook_node(self, nb, resources=None, **kw):
         """Convert from a notebook node."""
