+++ share/templates/webhtml/conf.json
@@ -1,0 +1,6 @@
+{
+  "base_template": "lab",
+  "mimetypes": {
+    "text/html": true
+  }
+}
