+++ share/templates/webhtml/index.html.j2
@@ -1,0 +1,1 @@
+{%- extends 'lab/index.html.j2' -%}
